+++ chat-master-server/chat-api/src/main/java/com/master/chat/api/config/InitBean.java
@@ -9,6 +9,7 @@ import com.master.chat.gpt.constant.BaseConfigConstant;
 import com.master.chat.gpt.mapper.OpenkeyMapper;
 import com.master.chat.gpt.pojo.vo.OpenkeyVO;
 import com.master.chat.llm.chatglm.ChatGLMClient;
+import com.master.chat.llm.doubao.DouBaoClient;
 import com.master.chat.llm.internlm.InternlmClient;
 import com.master.chat.llm.locallm.LocalLMClient;
 import com.master.chat.llm.moonshot.MoonshotClient;
@@ -240,6 +241,24 @@ public class InitBean {
     }
 
     /**
+     * 豆包
+     *
+     * @return
+     */
+    @Bean
+    public DouBaoClient douBaoClient() {
+        List<OpenkeyVO> openkeys = openkeyMapper.listOpenkeyByModel(ChatModelEnum.DOUBAO.getValue());
+        if (ValidatorUtil.isNullIncludeArray(openkeys)) {
+            log.error("未加载到豆包模型token数据，请添加后需要重启系统");
+            return new DouBaoClient();
+        }
+        if (ValidatorUtil.isNotNullIncludeArray(openkeys)) {
+            return DouBaoClient.builder().apiKey(openkeys.get(0).getAppKey()).build();
+        }
+        return DouBaoClient.builder().build();
+    }
+
+    /**
      * 书生·浦语
      *
      * @return
@@ -261,12 +280,16 @@ public class InitBean {
 
     /**
      * LocalLM 本地模型
-     * 支持Langchain-Chatchat、Ollama
+     * 支持Langchain-Chatchat、Ollama、GiteeAI、扣子、FastGPT
      *
      * @return
      */
     @Bean
     public LocalLMClient localLMClient() {
+        List<OpenkeyVO> openkeys = openkeyMapper.listOpenkeyByModel(ChatModelEnum.LOCALLM.getValue());
+        if (ValidatorUtil.isNotNullIncludeArray(openkeys)) {
+            return LocalLMClient.builder().apiKey(openkeys.get(0).getAppKey()).build();
+        }
         return LocalLMClient.builder().build();
     }