+++ chat-master-server/chat-gpt/src/main/java/com/master/chat/gpt/pojo/entity/Model.java
@@ -47,6 +47,11 @@ public class Model extends BaseEntity {
     private String version;
 
     /**
+     * 本地模型类型：1、Langchian；2、ollama；3、Giteeai 4、Coze
+     */
+    private Integer localModelType;
+
+    /**
      * 模型接口地址
      */
     private String modelUrl;
@@ -55,6 +60,11 @@ public class Model extends BaseEntity {
      * 知识库名称
      */
     private String knowledge;
+
+    /**
+     * 排序
+     */
+    private Integer sort;
 
     /**
      * 状态 0 禁用 1 启用