+++ chat-master-server/chat-llm/src/main/java/com/master/chat/llm/base/service/LLMService.java
@@ -16,8 +16,11 @@ import com.master.chat.llm.base.entity.ChatData;
 import com.master.chat.llm.base.exception.LLMException;
 import com.master.chat.llm.base.service.impl.*;
 import com.master.chat.llm.chatglm.ChatGLMClient;
+import com.master.chat.llm.doubao.DouBaoClient;
 import com.master.chat.llm.internlm.InternlmClient;
-import com.master.chat.llm.locallm.LocalLMClient;
+import com.master.chat.llm.locallm.coze.CozeClient;
+import com.master.chat.llm.locallm.langchain.LangchainClient;
+import com.master.chat.llm.locallm.ollama.OllamaClient;
 import com.master.chat.llm.moonshot.MoonshotClient;
 import com.master.chat.llm.openai.OpenAiClient;
 import com.master.chat.llm.openai.OpenAiStreamClient;
@@ -55,14 +58,17 @@ public class LLMService {
     private static TongYiClient tongYiClient;
     private static SparkClient sparkClient;
     private static MoonshotClient moonshotClient;
+    private static DouBaoClient douBaoClient;
     private static InternlmClient internlmClient;
-    private static LocalLMClient localLMClient;
+    private static LangchainClient langchainClient;
+    private static OllamaClient ollamaClient;
+    private static CozeClient cozeClient;
     private final GptService gptService;
 
     @Autowired
     public LLMService(GptService gptService, OpenAiClient openAiClient, OpenAiStreamClient openAiStreamClient, WenXinClient wenXinClient,
-                      ChatGLMClient chatGLMClient, TongYiClient tongYiClient, SparkClient sparkClient, MoonshotClient moonshotClient,
-                      InternlmClient internlmClient, LocalLMClient localLMClient) {
+                      ChatGLMClient chatGLMClient, TongYiClient tongYiClient, SparkClient sparkClient, MoonshotClient moonshotClient, DouBaoClient douBaoClient,
+                      InternlmClient internlmClient, LangchainClient langchainClient, OllamaClient ollamaClient, CozeClient cozeClient) {
         this.gptService = gptService;
         LLMService.openAiClient = openAiClient;
         LLMService.openAiStreamClient = openAiStreamClient;
@@ -71,8 +77,11 @@ public class LLMService {
         LLMService.tongYiClient = tongYiClient;
         LLMService.sparkClient = sparkClient;
         LLMService.moonshotClient = moonshotClient;
+        LLMService.douBaoClient = douBaoClient;
         LLMService.internlmClient = internlmClient;
-        LLMService.localLMClient = localLMClient;
+        LLMService.langchainClient = langchainClient;
+        LLMService.ollamaClient = ollamaClient;
+        LLMService.cozeClient = cozeClient;
     }
 
     public SseEmitter createSse(String uid) {
@@ -144,8 +153,10 @@ public class LLMService {
                 return new InternLMServiceImpl(internlmClient);
             case MOONSHOT:
                 return new MoonshotServiceImpl(moonshotClient);
+            case DOUBAO:
+                return new DouBaoServiceImpl(douBaoClient);
             case LOCALLM:
-                return new LocalLMServiceImpl(gptService,localLMClient);
+                return new LocalLMServiceImpl(langchainClient, ollamaClient, cozeClient, gptService);
             default:
                 return null;
         }
@@ -177,7 +188,7 @@ public class LLMService {
     }
 
 
-    public void sseChat(HttpServletResponse response, String uid, String conversationId) {
+    public void sseChat(HttpServletResponse response, Boolean isWs, String uid, String conversationId) {
         ChatMessageDTO chatMessage = gptService.getMessageByConverstationId(conversationId);
         String prompt = chatMessage.getContent();
         String version = chatMessage.getModelVersion();
@@ -195,7 +206,7 @@ public class LLMService {
         Integer status = ChatStatusEnum.SUCCESS.getValue();
         try {
             // ChatGPT、文心一言统一在SSEListener中处理流式返回，通义千问与讯飞星火\智谱清言单独处理
-            error = getLLM(modelEnum).streamChat(response, sseEmitter, chatMessages, isDraw(prompt), chatMessage.getChatId(), conversationId, prompt, version, uid);
+            error = getLLM(modelEnum).streamChat(response, sseEmitter, chatMessages, isWs, isDraw(prompt), chatMessage.getChatId(), conversationId, prompt, version, uid);
         } catch (BusinessException e) {
             throw e;
         } catch (Exception e) {