+++ chat-master-server/chat-llm/src/main/java/com/master/chat/llm/base/service/impl/LocalLMServiceImpl.java
@@ -5,23 +5,18 @@ import com.master.chat.client.model.command.ChatMessageCommand;
 import com.master.chat.client.model.dto.ChatMessageDTO;
 import com.master.chat.client.model.dto.ModelDTO;
 import com.master.chat.client.service.GptService;
-import com.master.chat.framework.validator.ValidatorUtil;
+import com.master.chat.common.exception.BusinessException;
 import com.master.chat.llm.base.service.ModelService;
-import com.master.chat.llm.locallm.LocalLMClient;
-import com.master.chat.llm.locallm.SSEListener;
-import com.master.chat.llm.locallm.chatchat.constant.ApiConstant;
-import com.master.chat.llm.locallm.chatchat.entity.ChatCompletion;
-import com.master.chat.llm.locallm.chatchat.entity.ChatCompletionMessage;
-import com.master.chat.llm.locallm.chatchat.enums.ModelEnum;
+import com.master.chat.llm.locallm.coze.CozeClient;
+import com.master.chat.llm.locallm.enums.ModelTypeEnum;
+import com.master.chat.llm.locallm.langchain.LangchainClient;
+import com.master.chat.llm.locallm.ollama.OllamaClient;
 import lombok.SneakyThrows;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
-import javax.annotation.Resource;
 import javax.servlet.http.HttpServletResponse;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 
 /**
@@ -35,15 +30,16 @@ import java.util.List;
  */
 @Service
 public class LocalLMServiceImpl implements ModelService {
-    private static LocalLMClient localLMClient;
-//    @Autowired
-//    private GptService gptService;
-
+    private static LangchainClient langchainClient;
+    private static OllamaClient ollamaClient;
+    private static CozeClient cozeClient;
     private static GptService gptService;
 
     @Autowired
-    public LocalLMServiceImpl(GptService gptService,LocalLMClient localLMClient) {
-        LocalLMServiceImpl.localLMClient = localLMClient;
+    public LocalLMServiceImpl(LangchainClient langchainClient, OllamaClient ollamaClient, CozeClient cozeClient, GptService gptService) {
+        LocalLMServiceImpl.langchainClient = langchainClient;
+        LocalLMServiceImpl.ollamaClient = ollamaClient;
+        LocalLMServiceImpl.cozeClient = cozeClient;
         LocalLMServiceImpl.gptService = gptService;
     }
 
@@ -54,32 +50,21 @@ public class LocalLMServiceImpl implements ModelService {
 
     @Override
     @SneakyThrows
-    public Boolean streamChat(HttpServletResponse response, SseEmitter sseEmitter, List<ChatMessageDTO> chatMessages, Boolean isDraw,
+    public Boolean streamChat(HttpServletResponse response, SseEmitter sseEmitter, List<ChatMessageDTO> chatMessages, Boolean isWs, Boolean isDraw,
                               Long chatId, String conversationId, String prompt, String version, String uid) {
-
-        List<ChatCompletionMessage> history = new ArrayList<>();
-        chatMessages.stream().forEach(v -> {
-            ChatCompletionMessage message = ChatCompletionMessage.builder().content(v.getContent()).role(v.getRole()).build();
-            history.add(message);
-        });
-        String modelVaersion = ValidatorUtil.isNotNull(version) ? version : ApiConstant.DEFAULT_MODEL;
-        SSEListener sseListener = new SSEListener(response, sseEmitter, chatId, conversationId, "langchain", version, prompt);
-        ChatCompletion chat = ChatCompletion
-                .builder()
-                .messages(chatMessages)
-//                .knowledgeBaseName(ApiConstant.DEFAULT_KNOWLEDGE)
-                .history(history)
-                .model(modelVaersion)
-                .presencePenalty(1.2)
-                .temperature(0.3)
-                .topP(0.7)
-                .maxTokens(8196)
-                .build();
-        ModelDTO model = gptService.getModel(ChatModelEnum.LOCALLM.getValue());
-        String domain = ValidatorUtil.isNotNull(model.getModelUrl()) ? model.getModelUrl() : ApiConstant.BASE_DOMAIN;
-        localLMClient.streamChat(chat, sseListener, ModelEnum.LLM, domain);
-        sseListener.getCountDownLatch().await();
-        return sseListener.getError();
+        ModelDTO modelDTO = gptService.getModel(ChatModelEnum.LOCALLM.getValue());
+        ModelTypeEnum modelType = ModelTypeEnum.getEnum(modelDTO.getLocalModelType());
+        switch (modelType) {
+            case LANGCHAIN:
+                return langchainClient.buildChatCompletion(response, sseEmitter, chatId, conversationId, isWs, uid, chatMessages, prompt, version, modelDTO);
+            case OLLAMA:
+                return ollamaClient.buildChatCompletion(response, sseEmitter, chatId, conversationId, isWs, uid, chatMessages, prompt, version, modelDTO);
+            case COZE:
+                return cozeClient.buildChatCompletion(response, sseEmitter, chatId, conversationId, isWs, uid, chatMessages, prompt, version, modelDTO);
+            default:
+                throw new BusinessException("未知的模型类型，功能未接入");
+        }
     }
+
 
 }